+++ Projects/App/Project.swift
@@ -14,12 +14,28 @@ let project = Project(
                 .spm(.FirebaseAnalyticsSwift),
                 .spm(.FirebaseDatabase),
                 .spm(.FirebaseDatabaseSwift),
-                .spm(.protoc),
+                //                .spm(.protoc),
                 .spm(.Realm),
                 .spm(.RealmSwift),
                 .spm(.SwiftProtobuf),
                 .spm(.SwiftProtobufPluginLibrary),
             ]
-        )
+        ),
+        .makeModule(
+            name: "\(Environment.appName)Tests",
+            product: .unitTests,
+            bundleId: "\(Environment.organizationName)Tests",
+            infoPlist: nil,
+            sources: ["Tests/Sources/**"],
+            resources: nil
+        ),
+        .makeModule(
+            name: "\(Environment.appName)UITests",
+            product: .uiTests,
+            bundleId: "\(Environment.organizationName)UITests",
+            infoPlist: nil,
+            sources: ["UITests/Sources/**"],
+            resources: nil
+        ),
     ]
 )
